Highlight trailing whitespace
Add rules directly to lexer.lua to highlight trailing whitespace on
lines with style ERROR for all lexers.  Add a rule before that to match
newlines so that the whitespace rule doesn't continue onto a new line
and consume whitespace there.  If it does then a line that is empty
except for whitespace isn't highlighted.